requirements.txt
tweepy==3.6.0
pycairo==1.17.0
